backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
psycopg2-binary==2.9.5
python-dotenv==1.0.0
bcrypt==4.0.1
pyjwt==2.6.0
email-validator==2.0.0
